pyproject.toml
[project]
name = "dynamo-sharepoint-ingest"
version = "0.1.0"
description = "SharePoint to S3 document ingestion pipeline with Textract extraction"
requires-python = ">=3.11"
dependencies = [
    "msal>=1.26.0",
    "requests>=2.31.0",
    "boto3>=1.34.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "moto>=5.0.0",
    "ruff>=0.2.0",
    "fpdf2>=2.7.0",
    "python-docx>=1.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "integration: integration tests requiring LocalStack",
]

[tool.ruff]
target-version = "py311"
line-length = 100
